Implement model saving and uploading functionality in ModelSystem; refactor activation method in model_training class.
